fix: move notification settings to settings file instead of using a different file
